Migrate SafeCallback into mCore
